qerer

+++ commands/clear.js
@@ -2,7 +2,7 @@ const Discord = require("discord.js");
 
 exports.run = async (bot, message, args) => {
 
-  //if(!message.member.hasPermission("MANAGE_MESSAGES")) return message.reply("**You do not have permission to do that.**");
+  if(!message.member.hasPermission("MANAGE_MESSAGES")) return message.reply("**You do not have permission to do that.**");
   if(!args[0]) return message.channel.send("Clear Command: **ium clear [amount of messages to clearn]**");
   message.channel.bulkDelete(args[0]).then(() => {
   message.delete().catch(O_o=>{});
